eems/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-autocomplete-light==3.8.1
django-bootstrap3==14.2.0
django-filter==2.4.0
django-tables2==2.3.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
